model-deployment/regression-model-k8s-deployment/requirements.txt
# Development and production requirements

# Core ML and data processing
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.3.0
joblib==1.3.0

# Web framework
Flask==2.3.0
Flask-CORS==4.0.0
gunicorn==21.0.0
Werkzeug==2.3.0

# MLflow and experiment tracking
mlflow==2.7.0
mlflow-skinny==2.7.0

# Configuration
PyYAML==6.0

# HTTP and networking
requests==2.31.0
urllib3==2.0.4

# Monitoring
prometheus-client==0.17.0

# Development and testing
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
isort==5.12.0
mypy==1.4.1

# Jupyter for notebooks (optional)
jupyter==1.0.0
jupyterlab==4.0.3
ipykernel==6.25.1

# Documentation
Sphinx==7.0.1
sphinx-rtd-theme==1.3.0